added ability to create user from API endpoint
