+++ app/controllers/better_together/bt/api/v1/communities_controller.rb
@@ -5,7 +5,7 @@ module BetterTogether
     module Api
       module V1   
         class CommunitiesController < ApiController
-          before_action :authenticate_better_together_user!, except: %i[index]
+          before_action :authenticate_user!, except: %i[index]
         end
       end
     end
