--- app/controllers/better_together/bt/api/v1/confirmations_controller.rb
@@ -1,17 +1,0 @@
-module BetterTogether
-  module Bt
-    module Api
-      module V1
-        class ConfirmationsController < Devise::ConfirmationsController
-          respond_to :json
-
-          protected
-
-          def resource_name
-            :user
-          end
-        end
-      end
-    end
-  end
-end
