--- app/controllers/better_together/bt/api/v1/sessions_controller.rb
@@ -1,25 +1,0 @@
-module BetterTogether
-  module Bt
-    module Api
-      module V1
-        class SessionsController < Devise::SessionsController
-          respond_to :json
-
-          protected
-
-          def respond_with(resource, _opts = {})
-            render json: resource
-          end
-
-          def respond_to_on_destroy
-            head :ok
-          end
-
-          def resource_name
-            :user
-          end
-        end
-      end
-    end
-  end
-end
