+++ app/controllers/better_together/confirmations_controller.rb
@@ -1,0 +1,11 @@
+module BetterTogether
+  class ConfirmationsController < Devise::ConfirmationsController
+    respond_to :json
+
+    protected
+
+    def resource_name
+      :user
+    end
+  end
+end
