+++ app/controllers/better_together/passwords_controller.rb
@@ -1,0 +1,11 @@
+module BetterTogether
+  class PasswordsController < Devise::PasswordsController
+    respond_to :json
+
+    protected
+
+    def resource_name
+      :user
+    end
+  end
+end
