+++ app/controllers/better_together/registrations_controller.rb
@@ -1,0 +1,12 @@
+module BetterTogether
+  class RegistrationsController < Devise::RegistrationsController
+    # before_action :configure_permitted_parameters
+    respond_to :json
+
+    protected
+
+    def resource_name
+      :user
+    end
+  end
+end
