+++ app/controllers/better_together/sessions_controller.rb
@@ -1,0 +1,19 @@
+module BetterTogether
+  class SessionsController < Devise::SessionsController
+    respond_to :json
+
+    protected
+
+    def respond_with(resource, _opts = {})
+      render json: resource
+    end
+
+    def respond_to_on_destroy
+      head :ok
+    end
+
+    def resource_name
+      :user
+    end
+  end
+end
