+++ spec/factories/better_together/users.rb
@@ -1,7 +1,9 @@
 require 'faker'
 
 FactoryBot.define do
-  factory :user do
+  factory :better_together_user,
+    class: ::BetterTogether::User,
+    aliases: %i[user] do
     email { Faker::Internet.unique.email }
     password { Faker::Internet.password(min_length: 10, max_length: 20) }
